fix: error safeproxy due to duplicate owners param
While creating or connecting a safe, issues arose with SafeProxy contract not deployed errors.
This was traced back to duplicate owner addresses being passed in the owners array.
To resolve this, the connected wallet address is now combined with user-entered owners and deduplicated before
being passed to the safe creation and connection functions.